Added papaparse to parse CSV files. Showing error messages when abbrivated statement limit is longer than 60 chars
