+++ composer.json
@@ -109,7 +109,6 @@
         "league/oauth2-github": "^0.2.2",
 
         "matthimatiker/opcache-bundle": "^0.4",
-        
         "bower-asset/datatables.net": "^1.10.15",
         "bower-asset/datatables.net-bs": "^1.10.15",
         "bower-asset/datatables.net-fixedheader": "^3.1.2",
@@ -120,7 +119,8 @@
         "bower-asset/datatables.net-select-bs": "^1.2.2",
         "liuggio/excelbundle": "^2.1",
         "bower-asset/jquery-comments": "^1.2",
-        "qandidate/toggle-bundle": "^0.8"
+        "qandidate/toggle-bundle": "^0.8",
+        "npm-asset/papaparse": "*"
     },
     "require-dev": {
         "sensio/generator-bundle": "^3.1",
